requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
inflection==0.5.1
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2024.2
uritemplate==4.1.1
